requirements.txt
asgiref==3.2.3
Django==3.0.2
python-slugify==4.0.0
pytz==2019.3
sqlparse==0.3.0
text-unidecode==1.3
